+++ deepcluster/benchmark.py
@@ -1,102 +1,202 @@
 import argparse
-from clustpy.data import load_mnist, load_cifar10
-from clustpy.deep import DEC, DCN, IDEC, VaDE
+from typing import Any
+
+from torch.utils.data import DataLoader
+
+from models.AlexNet import AlexNet
+from models.VGG import VGG16
+from models.FeedForward import FeedForward
+from deepcluster import DeepCluster
+from utils.kmeans import KMeans
+
+import torch
+import torchvision
+from torchvision.transforms import Compose, Resize, CenterCrop, ToTensor, Normalize, RandomResizedCrop, \
+    RandomHorizontalFlip
+import faiss
 import numpy as np
-from sklearn.metrics import normalized_mutual_info_score as nmi, silhouette_score as silhouette
-from clustpy.utils import EvaluationAlgorithm, evaluate_dataset, EvaluationMetric
+from torch.utils.data.sampler import RandomSampler
 import time
-import subprocess
-
-# utils -> evaluate
-
-DATASET_LOADERS = {
-    'mnist': load_mnist,
-    'cifar10': load_cifar10
-}
+import pandas as pd
 
 ALGORITHMS = {
-    'dec': DEC,
-    'dcn': DCN,
-    'vade': VaDE
+    'alexnet': AlexNet,
+    'vgg': VGG16,
+    'feedforward': FeedForward
+}
+
+DATASET = {
+    'cifar10': 'CIFAR10',
+    'mnist': 'MNIST',
 }
 
 
-def create_small_dataset(data, labels, sample_size=1000):
-    """Create a smaller dataset for testing."""
-    indices = np.random.choice(len(data), sample_size, replace=False)
-    return data[indices], labels[indices]
+def train_validation_data(data_dir: str, batch_size: int, seed: int, dataset: str = 'CIFAR10', valid_size: float = 0.1,
+                          shuffle=True) -> tuple:
+    print("Loading Dataset...")
+    if dataset == 'CIFAR10':
+        transform = Compose(
+            [
+                Resize(256),
+                CenterCrop(224),
+                ToTensor(),
+                Normalize(
+                    mean=[0.485, 0.456, 0.406],
+                    std=[0.229, 0.224, 0.225]
+                )
+            ]
+        )
+
+        train_data = torchvision.datasets.CIFAR10(
+            root=data_dir, train=True,
+            download=True, transform=transform,
+        )
+    elif dataset == 'MNIST':
+        transform = Compose(
+            [
+                Resize(256),
+                CenterCrop(224),
+                ToTensor(),
+                Normalize(
+                    (0.1307,), (0.3081,)
+                )
+            ]
+        )
+
+        train_data = torchvision.datasets.MNIST(
+            root=data_dir, train=True,
+            download=True, transform=transform,
+        )
+
+    print("Done Loading Dataset.")
+
+    train_sampler = RandomSampler(train_data)
+
+    train_loader = torch.utils.data.DataLoader(
+        train_data, batch_size=batch_size, sampler=train_sampler)
+
+    return train_loader
 
 
 def main(args):
-    data_loader = DATASET_LOADERS[args.dataset]
-    data, labels = data_loader()
-
-    # testing purpose
-    data, labels = create_small_dataset(data, labels)
-    n_clusters = len(np.unique(labels))
-
-
-    # benchmarking clustering algorithm
-    algorithm1 = ALGORITHMS[args.algorithm1]
-    algorithm2 = ALGORITHMS[args.algorithm2]
-
-    n_repetitions = 2
-    aggregations = [np.mean, np.std, np.max]
-
-    algorithms = [
-        EvaluationAlgorithm(name=args.algorithm1, algorithm=algorithm1, params={"n_clusters": None}),
-        EvaluationAlgorithm(name=args.algorithm2, algorithm=algorithm2, params={"n_clusters": None})
-    ]
-
-    metrics = [EvaluationMetric(name="nmi", metric=nmi, params={"average_method": "geometric"}, use_gt=True),
-               EvaluationMetric(name="silhouette", metric=silhouette, use_gt=False)]
-
-    #df = evaluate_dataset(X=data, evaluation_algorithms=algorithms, evaluation_metrics=metrics, labels_true=labels,
-    #                      n_repetitions=n_repetitions, aggregation_functions=aggregations, random_state=1, add_n_clusters=True)
-
-    #df.to_excel('test.xlsx')
-
     start_time = time.time()
 
-    subprocess.run(['python', 'train_deep_cluster.py'])
+    device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
+    print("Train on device:", device)
 
+    algorithm = ALGORITHMS[args.algorithm]
+
+    if args.algorithm == 'alexnet':
+        model = algorithm(input_dim=2, num_classes=10, sobel=True).to(device)
+    elif args.algorithm == 'vgg':
+        model = algorithm(input_dim=2, num_classes=10, sobel=True).to(device)
+    elif args.algorithm == 'feedforward':
+        model = algorithm(input_dim=224, num_classes=10).to(device)
+
+    torch.nn.utils.clip_grad_norm_(model.parameters(), max_norm=1.0)
+
+    # Optimizer
+    optimizer = torch.optim.SGD(
+        filter(lambda x: x.requires_grad, model.parameters()),
+        lr=args.lr,
+        momentum=0.9,
+        weight_decay=10 ** -5,
+    )
+
+    # Optimizer_TL
+    optimizer_tl = torch.optim.SGD(
+        model.top_layer.parameters(),
+        lr=args.lr,
+        weight_decay=10 ** -5,
+    )
+
+    # Loss Criterion
+    loss = torch.nn.CrossEntropyLoss()
+
+    # Cluster Assign Transformation
+    if args.dataset == 'cifar10':
+        normalize = Normalize(
+            mean=[0.4914, 0.4822, 0.4465],
+            std=[0.2023, 0.1994, 0.2010]
+        )
+    elif args.dataset == 'mnist':
+        normalize = Normalize(
+            (0.1307,), (0.3081,)
+        )
+
+    ca_tf = Compose(
+        [
+            RandomResizedCrop(224),
+            RandomHorizontalFlip(),
+            ToTensor(),
+            normalize
+        ]
+    )
+
+    batch_size = args.batch_size
+
+    DC_model = DeepCluster(
+        model=model,
+        optim=optimizer,
+        loss_criterion=loss,
+        optim_tl=optimizer_tl,
+        k=args.k,
+        batch_size=batch_size,
+        verbose=True,
+        pca_reduction=3,
+        cluster_assign_tf=ca_tf,
+        epochs=args.epochs,
+        dataset_name=DATASET[args.dataset],
+        clustering_method='sklearn'
+    )
+
+    train_loader = train_validation_data(data_dir='./data', batch_size=batch_size, seed=1)
+
+    print("Starting Training...")
+    DC_model.fit(train_loader)
+    print(DC_model.train_losses)
+    print("Training Done.")
     end_time = time.time()
-    execution_time = end_time - start_time
-    print(f"Execution time: {execution_time} seconds")
 
-    # benchmarking classification
-    # model = algo.algorithm(n_clusters=n_clusters)
-    # model.fit(data)
-    # embeddings = model.predict(data)  # Ensure predict method returns embeddings
-    # pseudo_labels = model.fit_predict(data)
-    #
-    # # Split data for classification
-    # train_embeddings, test_embeddings, train_labels, test_labels = train_test_split(embeddings, labels, test_size=0.2,
-    #                                                                                 random_state=42)
-    #
-    # # Train classifier with pseudo-labels
-    # acc_pseudo, prec_pseudo, rec_pseudo, f1_pseudo = train_classifier_with_pseudo_labels(train_embeddings,
-    #                                                                                      pseudo_labels, test_embeddings,
-    #                                                                                      test_labels)
-    #
-    # # Train classifier with real labels
-    # acc_real, prec_real, rec_real, f1_real = train_classifier_with_real_labels(train_embeddings, train_labels,
-    #                                                                            test_embeddings, test_labels)
-    #
-    # print(
-    #     f"{algo.name} - Classifier with Pseudo-Labels: Accuracy={acc_pseudo}, Precision={prec_pseudo}, Recall={rec_pseudo}, F1-Score={f1_pseudo}")
-    # print(
-    #     f"Classifier with Real Labels: Accuracy={acc_real}, Precision={prec_real}, Recall={rec_real}, F1-Score={f1_real}")
+    # save metrics
+    execution_time = end_time - start_time
+    losses = DC_model.train_losses
+    accuracies = DC_model.train_accuracies
+    classes = list(DC_model.train_classes)
+    nmi = DC_model.train_nmi
+    nmi.insert(0, 0)
+
+    metrics = {
+        'Epochs': list(range(1, args.epochs + 1)),
+        'Dataset': [args.dataset] * args.epochs,
+        'architecture': [args.algorithm] * args.epochs,
+        'Execution Time (s)': [execution_time] * args.epochs,
+        'Losses': losses,
+        'Accuracies': accuracies,
+        'Classes': [classes] * args.epochs,
+        'Nmi': [nmi]
+    }
+
+    df = pd.DataFrame(metrics)
+
+    print(df)
+    df.to_csv('training_metrics.csv', index=False)
 
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(description='Benchmark deep clustering algorithms.')
-    parser.add_argument('--dataset', type=str, choices=DATASET_LOADERS.keys(), required=True,
+    parser.add_argument('--dataset', type=str, choices=DATASET.keys(), required=True,
                         help='The dataset to use for benchmarking.')
-    parser.add_argument('--algorithm1', type=str, choices=ALGORITHMS.keys(), required=True,
-                        help='The algorithm which we compare with.')
-    parser.add_argument('--algorithm2', type=str, choices=ALGORITHMS.keys(), required=True,
-                        help='The algorithm which we compare with.')
+    parser.add_argument('--algorithm', type=str, choices=ALGORITHMS.keys(), required=True,
+                        help='The algorithm for training.')
+    parser.add_argument('--epochs', type=int, required=True,
+                        help='Amount of epochs.')
+    parser.add_argument('--lr', type=float, required=True,
+                        help='Learning Rate.')
+    parser.add_argument('--batch_size', type=int, required=True,
+                        help='Batch Size.')
+    parser.add_argument('--k', type=int, required=True,
+                        help='Amount of clusters.')
 
     args = parser.parse_args()
     main(args)
