--- deepcluster/test2.py
@@ -1,10 +1,0 @@
-import pandas as pd
-
-
-def main():
-    df = pd.read_csv('training_metrics.csv')
-    df.head()
-
-
-if __name__ == '__main__':
-    main()
